added middleware errorHandler
